If Overlapping of routes are detected, wildcards are expanded and is calculated again.
Routing includes the inventory to do an expansion in case it's needed.
__arc2DS maps from an Arclink address to a DataSelect service.
getRoute returns a tuple with a code and a result. There is a basic checking of overlap between the matching routes, to see whether the user can merge the results from different datacenters.
makeQueryGET checks the return code from getRoute. If there is overlapping (code:2) it expands the wildcards and calls getRoute again, so that the results can be safely merged.
If there are no wildcards and the expansion is empty, return and empty list.