Better algorithm for DataSelect. Output based on routes and NOT input parameters.
Previously, if all the request could be derived to one data center, the NSLC
parameters were the same as teh ones in the input. Now those are formed exactly
from the route entries in the routing table. This approach allows the data
centers to "hide" some streams if they declare the routes explicitly. F.i. the
fact that all the streams are matched by the input parameters do not allow me to
return exactly the input parameters.