Improvements in the information update procedure.
Move "updateTime" from the Global to the Service section in the configuration
file. "updTimes" is now a list of datetimes. There are attributes "lastUpd"
and "nextUpd" that keep track of the last update done and an index pointing to
the next update time in "updTimes".
A new method "updateAll" calls the three sources of information: "update",
"updateMT" and an InventoryCache object.
The first thing that "getRoute" does is to check if it is time to update. If
the time between the last update end now is more than 24 hours, an update is
mandatory.