Import routes from remote servers (synchronization!).
Service.synchronize is read from the configuration file. The format is two
elements per line separated by a comma, being the first one the name of the
datacenter and the second the base URL for the Routing Service.
For instance:
synchronize = SERVER1, URL1
    SERVER2, URL2
    SERVER3, URL3

No more Arclink configuration (now in updateAll.py).

Parsing of XML with a file handle and not a filename. Better for multiple
instances of a Python interpreter running in the same machine.