MAJOR CHANGE! Lookup is now based on priority instead of number of wildcards.
All tests have been run successfully, but this still needs to be tested with
more tests. In particular with network 4C.
The change was driven for the actual configuration of 4C and the fact that
overlaps are not being accepted per default. When the routes are processed
in incoming order, there are a couple of routes that are included until
4C.*.*.* is read. This is the last Route included and it does NOT overlap
because it has priority=1 (!!!) and the others priority=2. This is a great
test case.