Bugfix: Remote routes were not imported automatically (only by updateAll)
Now the system imports the routes already downloaded (routing-DCID.xml)
after importing the lcoal ones (routing.xml). Previously, it also worked
because updateAll did it and saved it in the binary cache.
This bug can only be reproduced if the system has to biuld the cache by
itself, what usually is not the case.